db/requirements.txt
psycopg2
kafka-python
pyyaml
dotenv
requests